+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Removed references to Laravel's form function to reduce dependencies
 - Added ability to delete records
 - Added rich card for record
+- Added error page if record not found in show view, previously was throwing 404 error
 
 ### Fixed
 - Updated tests